requirements.txt
Mastodon.py
click
